+++ examples/rbac/.wundergraph/wundergraph.config.ts
@@ -1,11 +1,4 @@
-import {
-	Application,
-	authProviders,
-	configureWunderGraphApplication,
-	cors,
-	introspect,
-	templates,
-} from '@wundergraph/sdk';
+import { authProviders, configureWunderGraphApplication, cors, introspect, templates } from '@wundergraph/sdk';
 import server from './wundergraph.server';
 import { NextJsTemplate } from '@wundergraph/nextjs/dist/template';
 import operations from './wundergraph.operations';
@@ -26,14 +19,9 @@ const spaceX = introspect.graphql({
 	url: 'https://spacex-api.fly.dev/graphql/',
 });
 
-const myApplication = new Application({
-	name: 'api',
-	apis: [spaceX, github],
-});
-
 // configureWunderGraph emits the configuration
 configureWunderGraphApplication({
-	application: myApplication,
+	apis: [spaceX, github],
 	server,
 	operations,
 	codeGenerators: [