+++ packages/sdk/src/codegen/index.test.ts
@@ -117,7 +117,6 @@ export const RunTemplateTest = async (...templates: Template[]) => {
 					},
 				},
 				application: {
-					Name: 'Test',
 					EngineConfiguration: new Api<any>('', [], [], [], []),
 					EnableSingleFlight: true,
 					S3UploadProvider: [],
@@ -433,6 +432,7 @@ export const RunTemplateTest = async (...templates: Template[]) => {
 							Internal: false,
 						},
 					],
+					InvalidOperationNames: [],
 					CorsConfiguration: {
 						maxAge: 120,
 						exposedHeaders: ['*'],
@@ -443,17 +443,13 @@ export const RunTemplateTest = async (...templates: Template[]) => {
 					},
 				},
 				deployment: {
-					path: 'api/main',
 					api: {
 						id: '',
-						name: 'api',
 					},
 					environment: {
-						name: 'main',
 						baseUrl: 'http://localhost:9991',
 						id: '',
 					},
-					name: 'main',
 				},
 				authentication: {
 					cookieBased: [],