+++ packages/sdk/src/definition/merge.test.ts
@@ -49,6 +49,7 @@ test('Should be merged', () => {
 					HooksConfiguration: {
 						onWSTransportConnectionInit: false,
 					},
+					CustomScalarTypeFields: [],
 				},
 				Directives: [],
 				RequestTimeoutSeconds: 0,
@@ -103,6 +104,7 @@ test('Should be merged', () => {
 					HooksConfiguration: {
 						onWSTransportConnectionInit: false,
 					},
+					CustomScalarTypeFields: [],
 				},
 				Directives: [],
 				RequestTimeoutSeconds: 0,
@@ -186,6 +188,7 @@ test('Should be merged', () => {
 					HooksConfiguration: {
 						onWSTransportConnectionInit: false,
 					},
+					CustomScalarTypeFields: [],
 				},
 				Directives: [],
 				RequestTimeoutSeconds: 0,
@@ -259,6 +262,7 @@ test('Should be merged', () => {
 					HooksConfiguration: {
 						onWSTransportConnectionInit: false,
 					},
+					CustomScalarTypeFields: [],
 				},
 				Directives: [],
 				RequestTimeoutSeconds: 0,
@@ -411,7 +415,7 @@ test('Should collide because weather and countries API has an enum called Langua
 						UseSSE: false,
 					},
 					Fetch: {
-						url: mapInputVariable('https://graphql-weather-api.herokuapp.com'),
+						url: mapInputVariable('https://weather-api.wundergraph.com'),
 						baseUrl: mapInputVariable(''),
 						path: mapInputVariable(''),
 						method: HTTPMethod.POST,
@@ -427,6 +431,7 @@ test('Should collide because weather and countries API has an enum called Langua
 					HooksConfiguration: {
 						onWSTransportConnectionInit: false,
 					},
+					CustomScalarTypeFields: [],
 				},
 				Directives: [],
 				RequestTimeoutSeconds: 0,