+++ packages/sdk/src/index.ts
@@ -64,7 +64,7 @@ export type {
 export type { WebhookVerifierConfiguration } from './webhooks/verifiers';
 
 export { default as templates } from './codegen/templates';
-export { Application, introspect, createMockApi, Api } from './definition';
+export { introspect, createMockApi, Api } from './definition';
 export { configureWunderGraphApplication, configurePublishWunderGraphAPI, resolveIntegration } from './configure';
 export { configureWunderGraphOperations, enableAuth, enableCaching, disableAuth } from './configure/operations';
 export { configureWunderGraphServer } from './middleware/server';
@@ -81,6 +81,7 @@ export {
 	TypeScriptResponseDataModels,
 	TypeScriptResponseModels,
 } from './codegen/templates/typescript/index';
+export { TypeScriptClient } from './codegen/templates/typescript/client';
 export { GithubWebhookVerifier, CreateWebhookVerifier, WebhookVerifierKind } from './webhooks/verifiers';
 export { PlaceHolder } from './configure/variables';
 export { EnvironmentVariable } from './configure/variables';